Fix parsing problems

+++ src/main/java/fr/gette/hciexplorer/entity/BeginRawMessage.java
@@ -2,6 +2,6 @@ package fr.gette.hciexplorer.entity;
 
 public interface BeginRawMessage extends RawMessage
 {
-	byte[] getInputBuffer();
-	void setInputBuffer(byte[] inputBuffer);
+	short[] getInputBuffer();
+	void setInputBuffer(short[] inputBuffer);
 }
